python-resume-parser/requirements.txt
# Complete requirements for Resume Parser & ATS Scoring System
# Install with: pip install -r requirements.txt

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.29.0

# Async Database & ORM
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.13.0

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# File Upload & Processing
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

# PDF Processing
pdfminer.six==20231228
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.1.0

# NLP & Machine Learning
scikit-learn==1.3.2
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Data Processing
numpy==1.24.4
pandas==2.1.4

# Text Processing
regex==2023.10.3
unidecode==1.3.7
python-dateutil==2.8.2

# Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
altair==5.2.0

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Database (Production)
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1